Resolved issues identified in code review
