+++ app/GlobalRedux/Feautures/category/category-types.ts
@@ -1,0 +1,3 @@
+export interface CategoryState {
+  categoryState: string;
+}
