+++ app/layout.tsx
@@ -1,6 +1,5 @@
-'use client';
-
 import { Inter } from 'next/font/google';
+import { Metadata } from 'next';
 import './globals.css';
 import Navbar from '../components/Navbar';
 import Footer from '../components/Footer';
@@ -9,10 +8,11 @@ import ReduxProvider from '../app/GlobalRedux/redux-provider';
 
 const inter = Inter({ subsets: ['latin'] });
 
-// export const metadata: Metadata = {
-//   title: 'Paws and Hearts',
-//   description: 'A animal shielter aims to match street dogs and cats with suitable keeper.',
-// }
+export const metadata: Metadata = {
+  title: 'Paws and Hearts',
+  description:
+    'A animal shielter aims to match street dogs and cats with suitable keeper.',
+};
 
 export default function RootLayout({
   children,
